Slight change to snapshot currency assumption
Allow use of snapshot when it's assumed current, and now also if the expected revision is either no provided or it's less than or equal to the snapshot.
It doesn't make sense to not use snapshot just because an expected revision is not provided OR if it's less than the snapshot.